Medic address from cep

+++ src/medics/medics.service.ts
@@ -1,22 +1,25 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException, ConflictException, NotFoundException } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
+import { AxiosResponse } from 'axios';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Repository, Like } from 'typeorm';
+
+import { map } from 'rxjs/operators';
 
 import { Medic } from './medics.entity';
 
 import { CreateMedicDto } from './medics.create.dto';
 import { UpdateMedicDto } from './medics.update.dto';
+import { AddressDto } from './medics.address.dto';
 
 @Injectable()
 export class MedicsService {
 
 	constructor(
 		@InjectRepository(Medic)
-		private medicsRepository: Repository<Medic>
+		private medicsRepository: Repository<Medic>,
+		private httpService: HttpService
 	) {}
-
-	// TODO: Check if medic specialties match the possile values
-	// TODO: Check if medic does not already exist
 
 	async exists(id: string): Promise<boolean> {
 
@@ -26,32 +29,76 @@ export class MedicsService {
 
 	}
 
-	create(medic: CreateMedicDto): object {
-		return this.medicsRepository.save(medic);
+	async addressFromCep(cep: string): Promise<AddressDto> {
+
+		return this.httpService
+			.get( `https://viacep.com.br/ws/${cep}/json`,)
+			.pipe(
+				map((res: AxiosResponse<any, any>) => {
+
+					const data = res.data;
+
+					console.log(data);
+
+					if(!data.cep)
+						throw new BadRequestException(`Invalid CEP: ${cep}`);
+
+					return {
+						street: data.logradouro,
+						area: data.bairro,
+						city: data.localidade,
+						state: data.uf,
+						complement: data.complemento, 
+					}
+
+				})
+			)
+			.toPromise();
+
+	}
+
+	async create(medic: CreateMedicDto): Promise<object> {
+
+		if(await this.exists(medic.crm))
+			throw new ConflictException(`Medic with CRM ${medic.crm} already exists`);
+
+		const address = await this.addressFromCep(medic.cep); 
+		console.log(address);
+		return this.medicsRepository.save({...medic, ...address});
+
 	}
 
 	readAll(query: any): Promise<Medic[]> {
-		return this.medicsRepository.find();
-	}
 
-	// TODO: Check if object exists on readOne, update and delete
-	// TODO: Make a pipe to check this
-	// TODO: Figure out how to send the error to the client 
+		return this.medicsRepository.find();
+
+	}
 
 	async readOne(id: string): Promise<Medic> {
+
 		if(!await this.exists(id))
-			throw `Medic with CRM ${id} does not exist`;
+			throw new NotFoundException(`Medic with CRM ${id} does not exist`)
+
 		return this.medicsRepository.findOneBy({ crm: id });
+
 	}
 
-	update(id: string, medic: UpdateMedicDto): object {
+	async update(id: string, medic: UpdateMedicDto): Promise<object> {
+
+		if(!await this.exists(id))
+			throw new NotFoundException(`Medic with CRM ${id} does not exist`)
 
 		return this.medicsRepository.update(id, medic);
 
 	}
 
 	async softDelete(id: string): Promise<void> {
+
+		if(!await this.exists(id))
+			throw new NotFoundException(`Medic with CRM ${id} does not exist`)
+
 		await this.medicsRepository.delete(id);
+
 	}
 
 }
